requirements.txt
fastapi
uvicorn
gitpython
pygithub